refactor(mcp): rename mcp_v2 to mcp_server
The `_v2` suffix in this codebase marks apps carried through the v0.93.0
v1->v2 multitenancy data migration (see backend/migrating/v2/README.md), not
current naming convention. Every app added since — pg_queue, dashboard_metrics,
platform_api, configuration — is unsuffixed. A brand-new app with no v1
predecessor should not claim the marker.

Named `mcp_server` rather than plain `mcp` because `mcp` is also the PyPI
package name for the MCP SDK (which mfbt's server uses). A bare `mcp` app
shadows it on the import path: `import mcp` resolved to the Django app, so a
later `pip install mcp` would break silently.
