+++ backend/mcp_server/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MCPConfig(AppConfig):
-    name = "mcp_v2"
+    name = "mcp_server"
